pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mineagent"
version = "0.0.1"
authors = [
    { name="Thomas Hopkins", email="thomas.hopkins.research@gmail.com" }
]
description = "Research on virtual intelligence in Minecraft"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Unix",
    "Development Status :: 1 - Planning",
]
dependencies = [
    "PyYAML",
    "dacite",
    "gymnasium",
    "numpy==1.26.4",
    "scipy",
    "torch",
    "torchvision",
    "tqdm",
    "tensorboard",
]

[tool.hatch.build.targets.wheel]
packages = ["mineagent"]

[project.license]
file = "LICENSE"

[project.urls]
Homepage = "https://github.com/thomashopkins32/Minecraft-Virtual-Intelligence"
Issues = "https://github.com/thomashopkins32/Minecraft-Virtual-Intelligence/issues"

[project.scripts]
mineagent = "mineagent.engine:run"

[project.optional-dependencies]
dev = [
    "pre-commit",
    "pyright",
    "pytest",
    "pytest-mock",
    "ruff",
    "types-PyYAML",
    "matplotlib",
]

[tool.pyright]
typeCheckingMode = "standard"
reportMissingImports = true
exclude = ["**/build"]

[tool.pixi.project]
channels = ["conda-forge", "pytorch"]
platforms = ["linux-64"]

[tool.pixi.host-dependencies]
hatchling = "*"

[tool.pixi.dependencies]
pytorch = "*"
torchvision = "*"
numpy = "==1.26.4"
scipy = "*"
tensorboard = "*"
gradle = "*"
openjdk = "21.*"


[tool.pixi.pypi-dependencies]
mineagent = { path = ".", editable = true }
MineDojo = { git = "https://github.com/thomashopkins32/MineDojo.git", rev = "main" }

[tool.pixi.environments]
default = { solve-group = "default" }
dev = { features = ["dev"], solve-group = "default" }
py310 = { features = ["py310"] }
py311 = { features = ["py311"] }
py312 = { features = ["py312"] }

[tool.pixi.feature.py310.dependencies]
python = "3.10.*"

[tool.pixi.feature.py311.dependencies]
python = "3.11.*"

[tool.pixi.feature.py312.dependencies]
python = "3.12.*"

[tool.pixi.tasks]
mineagent = "mineagent"
minecraft-client = "cd forge && gradle runClient"
minecraft-build = "cd forge && gradle build"
